CommentReaction, ArticleReaction and IReaction moved to entities.reactions package
