+++ src/main/java/org/blbulyandavbulyan/blog/repositories/reactions/CommentReactionRepository.java
@@ -1,7 +1,7 @@
 package org.blbulyandavbulyan.blog.repositories.reactions;
 
 import org.blbulyandavbulyan.blog.entities.Comment;
-import org.blbulyandavbulyan.blog.entities.CommentReaction;
+import org.blbulyandavbulyan.blog.entities.reactions.CommentReaction;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CommentReactionRepository extends JpaRepository<CommentReaction, Long>, IReactionRepository<CommentReaction, Comment>{
